Revert "feat(morello-fvp): add DPU node"
This reverts commit c624381044ca584bf9a4ca566ee6db126011e1b4.
The out-of-tree driver for the virtual encoder ("drm,virtual-encoder")
will no longer be carried in the Morello Linux kernel tree.
This also means that the mali-d71 will not be useable.

If the user want to utilize the display on the FVP the only supported
option is now the UEFI GOP.

Signed-off-by: Patrik Berglund <patrik.berglund@arm.com>
Change-Id: I671750f1581193c567090f30f5307164ebc32768